pkg/destroy/aws: Skip direct tag deletion for network interfaces
We create a lot of these in a given 24 hour period, because we try to
delete them before their consumer has been reaped:

  $ LOG="$(curl -s https://storage.googleapis.com/origin-ci-test/logs/release-openshift-origin-installer-e2e-aws-4.1/509/artifacts/e2e-aws/installer/.openshift_install.log | grep InvalidParameterValue)"
  $ echo "${LOG}" | wc -l
  75
  $ echo "${LOG}" | grep -v 'Network interface'
  ...no matches...

This will cool us off a bit, partly by limiting us to one ByVPC
failure in a given destroy cycle, vs. our current:

  $ echo "${LOG}" | head -n3
  time="2019-05-02T21:31:18Z" level=debug msg="deleting EC2 network interface eni-03b0e0c93880c0640: InvalidParameterValue: Network interface 'eni-03b0e0c93880c0640' is currently in use.\n\tstatus code: 400, request id: f59b52be-a091-4900-885b-34a42e221b89" arn="arn:aws:ec2:us-east-1:460538899914:vpc/vpc-0ba63aa6f96639ecd" id=vpc-0ba63aa6f96639ecd
  time="2019-05-02T21:31:18Z" level=debug msg="deleting EC2 network interface eni-0a4352bf3a63ebb1e: InvalidParameterValue: Network interface 'eni-0a4352bf3a63ebb1e' is currently in use.\n\tstatus code: 400, request id: e0ccfc6c-7a70-4013-8e7b-551700c1133c" arn="arn:aws:ec2:us-east-1:460538899914:vpc/vpc-0ba63aa6f96639ecd" id=vpc-0ba63aa6f96639ecd
  time="2019-05-02T21:31:18Z" level=debug msg="deleting EC2 network interface eni-02940af1a64eddc82: InvalidParameterValue: Network interface 'eni-02940af1a64eddc82' is currently in use.\n\tstatus code: 400, request id: 4c814c82-254e-4248-9848-30dad13ec66d" arn="arn:aws:ec2:us-east-1:460538899914:vpc/vpc-0ba63aa6f96639ecd" id=vpc-0ba63aa6f96639ecd
  $ echo "${LOG}" | grep vpc/vpc | wc -l
  53

With this change, that will drop to ~5:

  $ echo "${LOG}" | grep 'eni-02957f7396d3bfc5a.*vpc/vpc' | wc -l
  5

I'm also dropping the direct-tag removals completely, which will save:

  $ echo "${LOG}" | grep -v vpc/vpc | head -n1
  time="2019-05-02T21:31:21Z" level=debug msg="deleting arn:aws:ec2:us-east-1:460538899914:network-interface/eni-0a4352bf3a63ebb1e: InvalidParameterValue: Network interface 'eni-0a4352bf3a63ebb1e' is currently in use.\n\tstatus code: 400, request id: bd442d9a-ae6e-488c-bfb4-512821f5266c"
  $ echo "${LOG}" | grep -v vpc/vpc | wc -l
  22

And even if we removed all of these attempts, it seems AWS is cleaning
up the network interfaces when we remove the associated instance or
other consumer anyway.  The ByVPC method backstops any network
interfaces that happen to not be associated with an instance, load
balancer, or other resource.

By dropping direct-tag deletion, there is a change that we will leak
network interfaces created outside of tagged VPCs.  The installer will
certainly not set anything like that up, although it might happend in
user-provided infrastructure (UPI) flows where the VPC does not have a
matching tag but some network interfaces are still created and tagged
into the cluster.  At the moment, our AWS throttling issues seem like
a bigger worry than potential UPI leaks.

For the other resources whose ByVPC signature I'm touching with the
failFast addition, a slightly sleepier removal process is good because
resources of one type are likely to become unblocked for removal
around the same time, so a failure to remove one suggests the others
will not be ready for removal either.

+++ pkg/destroy/aws/aws.go
@@ -150,6 +150,11 @@ func (o *ClusterUninstaller) Run() error {
 										continue
 									}
 
+									if parsed.Service == "ec2" && strings.HasPrefix(parsed.Resource, "network-interface/") {
+										arnLogger.Debug("skip direct tag deletion, deleting the VPC should catch this")
+										continue
+									}
+
 									err = deleteARN(awsSession, parsed, filter, arnLogger)
 									if err != nil {
 										arnLogger.Debug(err)
@@ -714,7 +719,7 @@ func deleteEC2NATGateway(client *ec2.EC2, id string, logger logrus.FieldLogger) 
 	return nil
 }
 
-func deleteEC2NATGatewaysByVPC(client *ec2.EC2, vpc string, logger logrus.FieldLogger) error {
+func deleteEC2NATGatewaysByVPC(client *ec2.EC2, vpc string, failFast bool, logger logrus.FieldLogger) error {
 	var lastError error
 	err := client.DescribeNatGatewaysPages(
 		&ec2.DescribeNatGatewaysInput{
@@ -733,6 +738,9 @@ func deleteEC2NATGatewaysByVPC(client *ec2.EC2, vpc string, logger logrus.FieldL
 						logger.Debug(err)
 					}
 					lastError = errors.Wrapf(err, "deleting EC2 NAT gateway %s", *gateway.NatGatewayId)
+					if failFast {
+						return false
+					}
 				}
 			}
 
@@ -801,7 +809,7 @@ func deleteEC2RouteTableObject(client *ec2.EC2, table *ec2.RouteTable, logger lo
 	return nil
 }
 
-func deleteEC2RouteTablesByVPC(client *ec2.EC2, vpc string, logger logrus.FieldLogger) error {
+func deleteEC2RouteTablesByVPC(client *ec2.EC2, vpc string, failFast bool, logger logrus.FieldLogger) error {
 	var lastError error
 	err := client.DescribeRouteTablesPages(
 		&ec2.DescribeRouteTablesInput{
@@ -820,6 +828,9 @@ func deleteEC2RouteTablesByVPC(client *ec2.EC2, vpc string, logger logrus.FieldL
 						logger.Debug(err)
 					}
 					lastError = errors.Wrapf(err, "deleting EC2 route table %s", *table.RouteTableId)
+					if failFast {
+						return false
+					}
 				}
 			}
 
@@ -912,7 +923,7 @@ func deleteEC2NetworkInterface(client *ec2.EC2, id string, logger logrus.FieldLo
 	return nil
 }
 
-func deleteEC2NetworkInterfaceByVPC(client *ec2.EC2, vpc string, logger logrus.FieldLogger) error {
+func deleteEC2NetworkInterfaceByVPC(client *ec2.EC2, vpc string, failFast bool, logger logrus.FieldLogger) error {
 	var lastError error
 	err := client.DescribeNetworkInterfacesPages(
 		&ec2.DescribeNetworkInterfacesInput{
@@ -931,6 +942,9 @@ func deleteEC2NetworkInterfaceByVPC(client *ec2.EC2, vpc string, logger logrus.F
 						logger.Debug(lastError)
 					}
 					lastError = errors.Wrapf(err, "deleting EC2 network interface %s", *networkInterface.NetworkInterfaceId)
+					if failFast {
+						return false
+					}
 				}
 			}
 
@@ -987,13 +1001,13 @@ func deleteEC2VPC(ec2Client *ec2.EC2, elbClient *elb.ELB, elbv2Client *elbv2.ELB
 		return v2lbError
 	}
 
-	for _, helper := range [](func(client *ec2.EC2, vpc string, logger logrus.FieldLogger) error){
+	for _, helper := range [](func(client *ec2.EC2, vpc string, failFast bool, logger logrus.FieldLogger) error){
 		deleteEC2NATGatewaysByVPC,      // not always tagged
 		deleteEC2NetworkInterfaceByVPC, // not always tagged
 		deleteEC2RouteTablesByVPC,      // not always tagged
 		deleteEC2VPCEndpointsByVPC,     // not taggable
 	} {
-		err := helper(ec2Client, id, logger)
+		err := helper(ec2Client, id, true, logger)
 		if err != nil {
 			return err
 		}
@@ -1022,7 +1036,7 @@ func deleteEC2VPCEndpoint(client *ec2.EC2, id string, logger logrus.FieldLogger)
 	return nil
 }
 
-func deleteEC2VPCEndpointsByVPC(client *ec2.EC2, vpc string, logger logrus.FieldLogger) error {
+func deleteEC2VPCEndpointsByVPC(client *ec2.EC2, vpc string, failFast bool, logger logrus.FieldLogger) error {
 	response, err := client.DescribeVpcEndpoints(&ec2.DescribeVpcEndpointsInput{
 		Filters: []*ec2.Filter{
 			{
